Add support for Reclaim Policy and Bucket-Wipe

+++ src/main/java/com/emc/ecs/servicebroker/service/EcsServiceInstanceService.java
@@ -3,8 +3,10 @@ package com.emc.ecs.servicebroker.service;
 import com.emc.ecs.servicebroker.EcsManagementClientException;
 import com.emc.ecs.servicebroker.model.PlanProxy;
 import com.emc.ecs.servicebroker.model.ServiceDefinitionProxy;
+import com.emc.ecs.servicebroker.repository.LastOperationSerializer;
 import com.emc.ecs.servicebroker.repository.ServiceInstance;
 import com.emc.ecs.servicebroker.repository.ServiceInstanceRepository;
+import com.emc.object.s3.S3Exception;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,7 +18,10 @@ import org.springframework.cloud.servicebroker.service.ServiceInstanceService;
 import org.springframework.stereotype.Service;
 import reactor.core.publisher.Mono;
 
+import java.io.IOException;
 import java.util.Map;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.CompletionException;
 
 import static java.lang.String.format;
 
@@ -30,6 +35,8 @@ public class EcsServiceInstanceService implements ServiceInstanceService {
     private static final String BUCKET = "bucket";
     private static final String SERVICE_TYPE = "service-type";
     private static final Logger LOG = LoggerFactory.getLogger(EcsServiceInstanceService.class);
+
+    private static final LastOperationSerializer SUCCEEDED_OPERATION = new LastOperationSerializer(OperationState.SUCCEEDED, "", false);
 
     @Autowired
     private EcsService ecs;
@@ -85,20 +92,45 @@ public class EcsServiceInstanceService implements ServiceInstanceService {
         String serviceDefinitionId = request.getServiceDefinitionId();
         try {
             ServiceDefinitionProxy service = ecs
-                    .lookupServiceDefinition(serviceDefinitionId);
+                .lookupServiceDefinition(serviceDefinitionId);
             InstanceWorkflow workflow = getWorkflow(service)
-                    .withDeleteRequest(request);
+                .withDeleteRequest(request);
 
-            LOG.info("deleting service instance");
-            workflow.delete(serviceInstanceId);
+            ServiceInstance instance = null;
+            try {
+                instance = repository.find(serviceInstanceId);
+                if (instance == null) {
+                    LOG.info("Instance {} not found, assuming already deleted", serviceInstanceId);
+                    return Mono.just(DeleteServiceInstanceResponse.builder()
+                        .build());
+                }
+            } catch(S3Exception e) {
+                LOG.info("Instance {} not found, assuming already deleted", serviceInstanceId);
+                return Mono.just(DeleteServiceInstanceResponse.builder()
+                    .build());
+            }
 
-            LOG.info("removing instance from repo");
-            repository.delete(serviceInstanceId);
+            LOG.info("deleting service instance {}", serviceInstanceId);
+            CompletableFuture future = workflow.delete(serviceInstanceId);
+            if (future != null) {
+                instance.setLastOperation(new LastOperationSerializer(OperationState.IN_PROGRESS, "Deleting", true));
+                repository.save(instance);
+
+                // Setup callback to handle asynchronous delete completion
+                future.handle((result, exception) -> {
+                        asyncDeleteCompleted(serviceInstanceId, (Throwable)exception);
+                        return null;
+                    });
+            } else {
+                LOG.info("removing instance {} from repo", serviceInstanceId);
+                repository.delete(serviceInstanceId);
+            }
 
             return Mono.just(DeleteServiceInstanceResponse.builder()
-                    .async(false)
-                    .build());
+                .async(future != null)
+                .build());
         } catch (Exception e) {
+            logger.error("Error Deleting",e);
             throw new ServiceBrokerException(e);
         }
     }
@@ -140,6 +172,36 @@ public class EcsServiceInstanceService implements ServiceInstanceService {
         }
     }
 
+    @Override
+    public Mono<GetLastServiceOperationResponse> getLastOperation(GetLastServiceOperationRequest request) {
+        try {
+            ServiceInstance instance = repository.find(request.getServiceInstanceId());
+
+            if (instance == null)
+                throw new ServiceInstanceDoesNotExistException(request.getServiceInstanceId());
+
+            // No stored operation, assume succeeded
+            LastOperationSerializer lastOperation = instance.getLastOperation();
+            if (lastOperation == null) {
+                lastOperation = SUCCEEDED_OPERATION;
+            }
+
+            // Possibly delete the repository record
+            if (lastOperation.isDeleteOperation() && lastOperation.getOperationState() == OperationState.SUCCEEDED) {
+                logger.info("Operation for {} completed successfully, deleting from repository", instance.getServiceInstanceId(), lastOperation.getOperationState());
+                repository.delete(instance.getServiceInstanceId());
+            }
+
+            return Mono.just(GetLastServiceOperationResponse.builder()
+                .deleteOperation(lastOperation.isDeleteOperation())
+                .description(lastOperation.getDescription())
+                .operationState(lastOperation.getOperationState())
+                .build());
+        } catch (IOException e) {
+            throw new ServiceBrokerException(e);
+        }
+    }
+
     private InstanceWorkflow getWorkflow(CreateServiceInstanceRequest createRequest)
             throws EcsManagementClientException {
         if (isRemoteConnection(createRequest))
@@ -167,11 +229,29 @@ public class EcsServiceInstanceService implements ServiceInstanceService {
         }
     }
 
-    @Override
-    public Mono<GetLastServiceOperationResponse> getLastOperation(
-            GetLastServiceOperationRequest request) {
-        return Mono.just(GetLastServiceOperationResponse.builder()
-                .operationState(OperationState.SUCCEEDED)
-                .build());
+    private void asyncDeleteCompleted(String instanceId, Throwable exception) {
+        try {
+            ServiceInstance instance = repository.find(instanceId);
+            if (instance == null) {
+                logger.error("Unable to find instance {} when delete completed async", instanceId);
+            }
+
+            if (exception == null) {
+                instance.setLastOperation(new LastOperationSerializer(OperationState.SUCCEEDED, "Delete Complete", true));
+            } else {
+                String errorMsg;
+                if (exception instanceof CompletionException && exception.getCause() != null) {
+                    errorMsg = exception.getCause().getMessage();
+                } else {
+                    errorMsg = exception.getMessage();
+                }
+
+                instance.setLastOperation(new LastOperationSerializer(OperationState.FAILED, errorMsg, true));
+            }
+
+            repository.save(instance);
+        } catch (IOException e) {
+            logger.error("Unable to find instance {} when delete completed async");
+        }
     }
 }
